Renamed service.py to base.py for GCP resources
