+++ ScoutSuite/providers/gcp/resources/cloudresourcemanager/base.py
