+++ ScoutSuite/providers/gcp/resources/cloudstorage/base.py
