+++ ScoutSuite/providers/gcp/resources/gce/base.py
