+++ ScoutSuite/providers/gcp/resources/iam/base.py
