+++ ScoutSuite/providers/gcp/resources/stackdriverlogging/base.py
